refactor: move Traits/Authenticatable into User
